+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Game from './Game'
+import Game from './component/Game'
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
